requirements.txt
beautifulsoup4==4.5.3
future==0.16.0
oauthlib==2.0.1
pkg-resources==0.0.0
python-twitter==3.2.1
requests==2.13.0
requests-oauthlib==0.8.0
